feat: create HomeScreenProvider and useMovieContext hook
